requirements.txt
playwright>=1.40.0
playwright-stealth>=2.0.0
pydantic>=2.5.0
rich>=13.7.0
gspread>=6.0.0
google-auth>=2.0.0
